% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\docType{data}
\name{kat_labels}
\alias{kat_labels}
\title{Digital theme category labels}
\format{
A named character vector of length 9. Names are the integer codes
as character strings (\code{"1"} to \code{"9"}); values are German labels.

\describe{
\item{1}{Infrastruktur}
\item{2}{Digitale Wirtschaft}
\item{3}{Digitale Verwaltung}
\item{4}{Digitale Kompetenzen}
\item{5}{Kultur & Zivilgesellschaft}
\item{6}{Forschung & Innovation}
\item{7}{Gesundheitswesen}
\item{8}{Bundeswehr}
\item{9}{Unteilbare Ausgaben}
}
}
\usage{
kat_labels
}
\description{
A named character vector mapping the nine digital theme category codes
(column \code{kategorie} in \link{digitalhaushalt}) to their German names.
}
\examples{
kat_labels["3"]
# "Digitale Verwaltung"

# Add readable labels to a subset of the data
data(digitalhaushalt)
digi <- subset(digitalhaushalt, !is.na(kategorie))
digi$kat_name <- kat_labels[as.character(digi$kategorie)]
}
\seealso{
\link{digitalhaushalt}, \link{hf_labels}, \link{of_labels}
}
\keyword{datasets}
